+++ redis_broadcast.go
@@ -113,6 +113,7 @@ func newRedisBroadcast(nsp string, adapter *RedisAdapterOptions) (*redisBroadcas
 	if err != nil {
 		return nil, err
 	}
+
 	sub, err := redis.Dial("tcp", redisAddr)
 	if err != nil {
 		return nil, err
@@ -157,6 +158,159 @@ func newRedisBroadcast(nsp string, adapter *RedisAdapterOptions) (*redisBroadcas
 	return &bc, nil
 }
 
+// AllRooms gives list of all rooms available for redisBroadcast.
+func (bc *redisBroadcast) AllRooms() []string {
+	req := allRoomRequest{
+		RequestType: allRoomReqType,
+		RequestID:   newV4UUID(),
+	}
+	reqJSON, _ := json.Marshal(&req)
+
+	req.rooms = make(map[string]bool)
+	numSub, _ := bc.getNumSub(bc.reqChannel)
+	req.numSub = numSub
+	req.done = make(chan bool, 1)
+
+	bc.requests[req.RequestID] = &req
+	bc.pub.Conn.Do("PUBLISH", bc.reqChannel, reqJSON)
+
+	<-req.done
+	rooms := make([]string, 0, len(req.rooms))
+	for room := range req.rooms {
+		rooms = append(rooms, room)
+	}
+
+	delete(bc.requests, req.RequestID)
+	return rooms
+}
+
+// Join joins the given connection to the redisBroadcast room.
+func (bc *redisBroadcast) Join(room string, connection Conn) {
+	bc.lock.Lock()
+	defer bc.lock.Unlock()
+
+	if _, ok := bc.rooms[room]; !ok {
+		bc.rooms[room] = make(map[string]Conn)
+	}
+
+	bc.rooms[room][connection.ID()] = connection
+}
+
+// Leave leaves the given connection from given room (if exist)
+func (bc *redisBroadcast) Leave(room string, connection Conn) {
+	bc.lock.Lock()
+	defer bc.lock.Unlock()
+
+	if connections, ok := bc.rooms[room]; ok {
+		delete(connections, connection.ID())
+
+		if len(connections) == 0 {
+			delete(bc.rooms, room)
+		}
+	}
+}
+
+// LeaveAll leaves the given connection from all rooms.
+func (bc *redisBroadcast) LeaveAll(connection Conn) {
+	bc.lock.Lock()
+	defer bc.lock.Unlock()
+
+	for room, connections := range bc.rooms {
+		delete(connections, connection.ID())
+
+		if len(connections) == 0 {
+			delete(bc.rooms, room)
+		}
+	}
+}
+
+// Clear clears the room.
+func (bc *redisBroadcast) Clear(room string) {
+	bc.lock.Lock()
+	defer bc.lock.Unlock()
+
+	delete(bc.rooms, room)
+	go bc.publishClear(room)
+}
+
+// Send sends given event & args to all the connections in the specified room.
+func (bc *redisBroadcast) Send(room, event string, args ...interface{}) {
+	bc.lock.RLock()
+	defer bc.lock.RUnlock()
+
+	connections, ok := bc.rooms[room]
+	if ok {
+		for _, connection := range connections {
+			connection.Emit(event, args...)
+		}
+	}
+
+	bc.publishMessage(room, event, args...)
+}
+
+// SendAll sends given event & args to all the connections to all the rooms.
+func (bc *redisBroadcast) SendAll(event string, args ...interface{}) {
+	bc.lock.RLock()
+	defer bc.lock.RUnlock()
+
+	for _, connections := range bc.rooms {
+		for _, connection := range connections {
+			connection.Emit(event, args...)
+		}
+	}
+	bc.publishMessage("", event, args...)
+}
+
+// ForEach sends data returned by DataFunc, if room does not exits sends nothing.
+func (bc *redisBroadcast) ForEach(room string, f EachFunc) {
+	bc.lock.RLock()
+	defer bc.lock.RUnlock()
+
+	occupants, ok := bc.rooms[room]
+	if !ok {
+		return
+	}
+
+	for _, connection := range occupants {
+		f(connection)
+	}
+}
+
+// Len gives number of connections in the room.
+func (bc *redisBroadcast) Len(room string) int {
+	req := roomLenRequest{
+		RequestType: roomLenReqType,
+		RequestID:   newV4UUID(),
+		Room:        room,
+	}
+
+	reqJSON, _ := json.Marshal(&req)
+	numSub, _ := bc.getNumSub(bc.reqChannel)
+	req.numSub = numSub
+	req.done = make(chan bool, 1)
+
+	bc.requests[req.RequestID] = &req
+	bc.pub.Conn.Do("PUBLISH", bc.reqChannel, reqJSON)
+	<-req.done
+
+	delete(bc.requests, req.RequestID)
+	return req.connections
+}
+
+// Rooms gives the list of all the rooms available for redisBroadcast in case of
+// no connection is given, in case of a connection is given, it gives
+// list of all the rooms the connection is joined to.
+func (bc *redisBroadcast) Rooms(connection Conn) []string {
+	bc.lock.RLock()
+	defer bc.lock.RUnlock()
+
+	if connection == nil {
+		return bc.AllRooms()
+	}
+
+	return bc.getRoomsByConn(connection)
+}
+
 func (bc *redisBroadcast) onMessage(channel string, msg []byte) error {
 	channelParts := strings.Split(channel, "#")
 	nsp := channelParts[len(channelParts)-2]
@@ -194,7 +348,7 @@ func (bc *redisBroadcast) onMessage(channel string, msg []byte) error {
 	return nil
 }
 
-// Get the number of subcribers of a channel
+// Get the number of subscribers of a channel.
 func (bc *redisBroadcast) getNumSub(channel string) (int, error) {
 	rs, err := bc.pub.Conn.Do("PUBSUB", "NUMSUB", channel)
 	if err != nil {
@@ -206,7 +360,7 @@ func (bc *redisBroadcast) getNumSub(channel string) (int, error) {
 	return int(numSub64), nil
 }
 
-// Handle request from redis channel
+// Handle request from redis channel.
 func (bc *redisBroadcast) onRequest(msg []byte) {
 	var req map[string]string
 	err := json.Unmarshal(msg, &req)
@@ -250,7 +404,7 @@ func (bc *redisBroadcast) publish(channel string, msg interface{}) {
 	bc.pub.Conn.Do("PUBLISH", channel, resJSON)
 }
 
-// Handle response from redis channel
+// Handle response from redis channel.
 func (bc *redisBroadcast) onResponse(msg []byte) {
 	var res map[string]interface{}
 	err := json.Unmarshal(msg, &res)
@@ -300,55 +454,6 @@ func (bc *redisBroadcast) onResponse(msg []byte) {
 	}
 }
 
-// Join joins the given connection to the redisBroadcast room
-func (bc *redisBroadcast) Join(room string, connection Conn) {
-	bc.lock.Lock()
-	defer bc.lock.Unlock()
-
-	if _, ok := bc.rooms[room]; !ok {
-		bc.rooms[room] = make(map[string]Conn)
-	}
-
-	bc.rooms[room][connection.ID()] = connection
-}
-
-// Leave leaves the given connection from given room (if exist)
-func (bc *redisBroadcast) Leave(room string, connection Conn) {
-	bc.lock.Lock()
-	defer bc.lock.Unlock()
-
-	if connections, ok := bc.rooms[room]; ok {
-		delete(connections, connection.ID())
-
-		if len(connections) == 0 {
-			delete(bc.rooms, room)
-		}
-	}
-}
-
-// LeaveAll leaves the given connection from all rooms
-func (bc *redisBroadcast) LeaveAll(connection Conn) {
-	bc.lock.Lock()
-	defer bc.lock.Unlock()
-
-	for room, connections := range bc.rooms {
-		delete(connections, connection.ID())
-
-		if len(connections) == 0 {
-			delete(bc.rooms, room)
-		}
-	}
-}
-
-// Clear clears the room
-func (bc *redisBroadcast) Clear(room string) {
-	bc.lock.Lock()
-	defer bc.lock.Unlock()
-
-	delete(bc.rooms, room)
-	go bc.publishClear(room)
-}
-
 func (bc *redisBroadcast) publishClear(room string) {
 	req := clearRoomRequest{
 		RequestType: clearRoomReqType,
@@ -365,21 +470,6 @@ func (bc *redisBroadcast) clear(room string) {
 	defer bc.lock.Unlock()
 
 	delete(bc.rooms, room)
-}
-
-// Send sends given event & args to all the connections in the specified room
-func (bc *redisBroadcast) Send(room, event string, args ...interface{}) {
-	bc.lock.RLock()
-	defer bc.lock.RUnlock()
-
-	connections, ok := bc.rooms[room]
-	if ok {
-		for _, connection := range connections {
-			connection.Emit(event, args...)
-		}
-	}
-
-	bc.publishMessage(room, event, args...)
 }
 
 func (bc *redisBroadcast) send(room string, event string, args ...interface{}) {
@@ -410,19 +500,6 @@ func (bc *redisBroadcast) publishMessage(room string, event string, args ...inte
 	bc.pub.Conn.Do("PUBLISH", bc.key, bcMessageJSON)
 }
 
-// SendAll sends given event & args to all the connections to all the rooms
-func (bc *redisBroadcast) SendAll(event string, args ...interface{}) {
-	bc.lock.RLock()
-	defer bc.lock.RUnlock()
-
-	for _, connections := range bc.rooms {
-		for _, connection := range connections {
-			connection.Emit(event, args...)
-		}
-	}
-	bc.publishMessage("", event, args...)
-}
-
 func (bc *redisBroadcast) sendAll(event string, args ...interface{}) {
 	bc.lock.RLock()
 	defer bc.lock.RUnlock()
@@ -432,82 +509,6 @@ func (bc *redisBroadcast) sendAll(event string, args ...interface{}) {
 			connection.Emit(event, args...)
 		}
 	}
-}
-
-// ForEach sends data returned by DataFunc, if room does not exits sends nothing
-func (bc *redisBroadcast) ForEach(room string, f EachFunc) {
-	bc.lock.RLock()
-	defer bc.lock.RUnlock()
-
-	occupants, ok := bc.rooms[room]
-	if !ok {
-		return
-	}
-
-	for _, connection := range occupants {
-		f(connection)
-	}
-}
-
-// Len gives number of connections in the room
-func (bc *redisBroadcast) Len(room string) int {
-	req := roomLenRequest{
-		RequestType: roomLenReqType,
-		RequestID:   newV4UUID(),
-		Room:        room,
-	}
-
-	reqJSON, _ := json.Marshal(&req)
-	numSub, _ := bc.getNumSub(bc.reqChannel)
-	req.numSub = numSub
-	req.done = make(chan bool, 1)
-
-	bc.requests[req.RequestID] = &req
-	bc.pub.Conn.Do("PUBLISH", bc.reqChannel, reqJSON)
-	<-req.done
-
-	delete(bc.requests, req.RequestID)
-	return req.connections
-}
-
-// Rooms gives the list of all the rooms available for redisBroadcast in case of
-// no connection is given, in case of a connection is given, it gives
-// list of all the rooms the connection is joined to
-func (bc *redisBroadcast) Rooms(connection Conn) []string {
-	bc.lock.RLock()
-	defer bc.lock.RUnlock()
-
-	if connection == nil {
-		return bc.AllRooms()
-	}
-
-	return bc.getRoomsByConn(connection)
-}
-
-// AllRooms gives list of all rooms available for redisBroadcast
-func (bc *redisBroadcast) AllRooms() []string {
-	req := allRoomRequest{
-		RequestType: allRoomReqType,
-		RequestID:   newV4UUID(),
-	}
-	reqJSON, _ := json.Marshal(&req)
-
-	req.rooms = make(map[string]bool)
-	numSub, _ := bc.getNumSub(bc.reqChannel)
-	req.numSub = numSub
-	req.done = make(chan bool, 1)
-
-	bc.requests[req.RequestID] = &req
-	bc.pub.Conn.Do("PUBLISH", bc.reqChannel, reqJSON)
-
-	<-req.done
-	rooms := make([]string, 0, len(req.rooms))
-	for room := range req.rooms {
-		rooms = append(rooms, room)
-	}
-
-	delete(bc.requests, req.RequestID)
-	return rooms
 }
 
 func (bc *redisBroadcast) allRooms() []string {
