+++ types.go
@@ -1,5 +1,10 @@
 package socketio
 
+import (
+	"github.com/googollee/go-socket.io/parser"
+	"reflect"
+)
+
 // namespace
 const (
 	aliasRootNamespace = "/"
@@ -9,4 +14,10 @@ const (
 // message
 const (
 	clientDisconnectMsg = "client namespace disconnect"
+)
+
+type readHandler func(c *conn, header parser.Header) error
+
+var (
+	defaultHeaderType = []reflect.Type{reflect.TypeOf("")}
 )
